+++ stroom-query-common/src/main/java/stroom/query/common/v2/SearchResponseCreatorCache.java
@@ -22,6 +22,10 @@ public interface SearchResponseCreatorCache {
 
     void evictExpiredElements();
 
+    /**
+     * Clears all items from the cache
+     */
+    void clear();
 
     //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~